Include session in unauthenticated admin return
