+++ packages/shopify-app-remix/src/server/__tests__/shopify-app.test.ts
@@ -12,10 +12,7 @@ import {
 } from '../index';
 import {AppConfigArg} from '../config-types';
 
-import {TEST_SHOP, setUpValidSession, testConfig} from './test-helper';
-import {mockExternalRequest} from './request-mock';
-
-const REQUEST_URL = `https://${TEST_SHOP}/admin/api/${shopifyApiPackage.LATEST_API_VERSION}/customers.json`;
+import {testConfig} from './test-helper';
 
 describe('shopifyApp', () => {
   /* eslint-disable no-process-env */
@@ -118,111 +115,5 @@ describe('shopifyApp', () => {
 
     // THEN
     expect(() => shopifyApp(config)).toThrowError(ShopifyError);
-  });
-
-  describe('unauthenticated admin context', () => {
-    it('REST client can perform GET requests', async () => {
-      // GIVEN
-      const shopify = shopifyApp(testConfig());
-      await setUpValidSession(shopify.sessionStorage, false);
-      const {admin} = await shopify.unauthenticated.admin(TEST_SHOP);
-
-      await mockExternalRequest({
-        request: new Request(REQUEST_URL),
-        response: new Response(JSON.stringify({customers: []})),
-      });
-
-      // WHEN
-      const response = await admin.rest.get({path: 'customers'});
-
-      // THEN
-      expect(response.status).toEqual(200);
-      expect(await response.json()).toEqual({customers: []});
-    });
-
-    it('REST client can perform POST requests', async () => {
-      // GIVEN
-      const shopify = shopifyApp(testConfig());
-      await setUpValidSession(shopify.sessionStorage, false);
-      const {admin} = await shopify.unauthenticated.admin(TEST_SHOP);
-
-      await mockExternalRequest({
-        request: new Request(REQUEST_URL, {method: 'POST'}),
-        response: new Response(JSON.stringify({customers: []})),
-      });
-
-      // WHEN
-      const response = await admin.rest.post({
-        path: '/customers.json',
-        data: '',
-      });
-
-      // THEN
-      expect(response.status).toEqual(200);
-      expect(await response.json()).toEqual({customers: []});
-    });
-
-    it('REST client can perform PUT requests', async () => {
-      // GIVEN
-      const shopify = shopifyApp(testConfig());
-      await setUpValidSession(shopify.sessionStorage, false);
-      const {admin} = await shopify.unauthenticated.admin(TEST_SHOP);
-
-      await mockExternalRequest({
-        request: new Request(REQUEST_URL, {method: 'PUT'}),
-        response: new Response(JSON.stringify({customers: []})),
-      });
-
-      // WHEN
-      const response = await admin.rest.put({
-        path: '/customers.json',
-        data: '',
-      });
-
-      // THEN
-      expect(response.status).toEqual(200);
-      expect(await response.json()).toEqual({customers: []});
-    });
-
-    it('REST client can perform DELETE requests', async () => {
-      // GIVEN
-      const shopify = shopifyApp(testConfig());
-      await setUpValidSession(shopify.sessionStorage, false);
-      const {admin} = await shopify.unauthenticated.admin(TEST_SHOP);
-
-      await mockExternalRequest({
-        request: new Request(REQUEST_URL, {method: 'DELETE'}),
-        response: new Response(JSON.stringify({customers: []})),
-      });
-
-      // WHEN
-      const response = await admin.rest.delete({path: '/customers.json'});
-
-      // THEN
-      expect(response.status).toEqual(200);
-      expect(await response.json()).toEqual({customers: []});
-    });
-
-    it('GraphQL client can perform requests', async () => {
-      // GIVEN
-      const shopify = shopifyApp(testConfig());
-      await setUpValidSession(shopify.sessionStorage, false);
-      const {admin} = await shopify.unauthenticated.admin(TEST_SHOP);
-
-      await mockExternalRequest({
-        request: new Request(
-          `https://${TEST_SHOP}/admin/api/${shopifyApiPackage.LATEST_API_VERSION}/graphql.json`,
-          {method: 'POST'},
-        ),
-        response: new Response(JSON.stringify({shop: {name: 'Test shop'}})),
-      });
-
-      // WHEN
-      const response = await admin.graphql('{ shop { name } }');
-
-      // THEN
-      expect(response.status).toEqual(200);
-      expect(await response.json()).toEqual({shop: {name: 'Test shop'}});
-    });
   });
 });
