+++ packages/shopify-app-remix/src/server/unauthenticated/admin/factory.ts
@@ -3,10 +3,14 @@ import {ShopifyError, ShopifyRestResources} from '@shopify/shopify-api';
 import {BasicParams} from '../../types';
 import {adminClientFactory} from '../../clients/admin';
 
+import {UnauthenticatedAdminContext} from './types';
+
 export function unauthenticatedAdminContextFactory<
   Resources extends ShopifyRestResources,
 >(params: BasicParams) {
-  return async (shop: string) => {
+  return async (
+    shop: string,
+  ): Promise<UnauthenticatedAdminContext<Resources>> => {
     const offlineSessionId = params.api.session.getOfflineId(shop);
     const session = await params.config.sessionStorage.loadSession(
       offlineSessionId,
@@ -19,6 +23,7 @@ export function unauthenticatedAdminContextFactory<
     }
 
     return {
+      session,
       admin: adminClientFactory<Resources>({params, session}),
     };
   };
